1、使用httpclient测试接口 auth.http、cart.http、merchandise.http、shop.http star.http、user.http、system.http 2、修改部分代码 1）redis的hashkey不能是Long类型，需要toString一下 2）加一个LocalDateTimeConfig用来做java8的时间类的转化 3）增加了几个参数校验，并自定义校验规则@Sex 4）Voucher中的一个字段limit需要加上``号 5）为feign修改白名单的放行规则 6）增加mp的分页插件 7）。。。 3、还剩下voucher.http、order.http

+++ qian-bian-handmade-shop-feign/src/main/java/com/lmm/entity/DeliveryAddress.java
@@ -21,7 +21,7 @@ public class DeliveryAddress implements Serializable {
 
     @ApiModelProperty("收货地址id")
     @TableId(value = "id", type = IdType.ASSIGN_ID)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty("用户编号")
     @TableField("user_id")
@@ -55,12 +55,11 @@ public class DeliveryAddress implements Serializable {
     @TableField("priority")
     private Integer priority;
 
-
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
